generalise failing promise tests
